backend/requirements.txt
fastapi
uvicorn[standard]
strawberry-graphql[fastapi]
pymongo
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
faker
jinja2
python-multipart
aiofiles
motor
PyJWT